+++ CHANGELOG.md
@@ -11,10 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added CMIP6 data source
+- Added CBottle Tropical Cyclone guidance diagnostic
+- Exposed backend arguments of netcdf/zarr to datasource_to_file signature
+- Added vertical wind speed support in GFS
 
 ### Changed
 
-- Added vertical wind speed support in GFS
+- Updated CBottle data source to mixture of experts models
 
 ### Deprecated
 
@@ -25,6 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 ### Dependencies
+
+- Dropped support for Python 3.10
+- Bumped CBottle (and Earth2Grid) versions
 
 ## [0.9.0] - 2025-08-19